better worker summary print; clean up tests

+++ binstar_build_client/tests/test_worker_script.py
@@ -16,9 +16,10 @@ import unittest
 from binstar_client.tests.fixture import CLITestCase
 from binstar_client.tests.urlmock import urlpatch
 from binstar_build_client.scripts.build import main
+from binstar_build_client.worker.register import REGISTERED_WORKERS_DIR
 
 worker_data = {'cwd': '.',
-               'output': os.path.join(tempfile.gettempdir(), 'worker.yaml'), 
+               'output': os.path.join(REGISTERED_WORKERS_DIR, 'worker_id'), 
                'username': 'username',
                'queue': 'queue-1',
                'conda_build_dir': 'conda_build_dir',
@@ -26,7 +27,11 @@ worker_data = {'cwd': '.',
                'token': 'token',
                'platform': 'platform',
                'worker_id': 'worker_id',
-               'status_file': 'status_file',}
+               'status_file': 'status_file',
+               'hostname': 'localhost',
+               'dist': 'dist',
+               'max_job_duration': 1000,
+            }
 
 class Test(CLITestCase):
 
@@ -51,9 +56,7 @@ class Test(CLITestCase):
     @patch('binstar_build_client.commands.deregister.deregister_worker')
     def test_deregister_from_id(self, urls, deregister_worker):
 
-        main(['deregister', '-q', 
-              '{}/{}'.format(worker_data['username'], worker_data['queue']),
-              '--worker-id', worker_data['worker_id']], False)
+        main(['deregister', '--worker-id', worker_data['worker_id']], False)
         self.assertEqual(deregister_worker.call_count, 1)
     
 
@@ -62,9 +65,13 @@ class Test(CLITestCase):
     def test_worker_simple(self, urls, Worker):
         with open(worker_data['output'], 'w') as f:
             f.write(yaml.dump(worker_data))
-        main(['--show-traceback', 'worker', worker_data['output']], False)
+        main(['--show-traceback', 'worker', worker_data['worker_id']], False)
         self.assertEqual(Worker().work_forever.call_count, 1)
-        os.remove(worker_data['output'])
+        
+    @classmethod
+    def tearDownClass(cls):
+        if os.path.exists(worker_data['output']):
+            os.remove(worker_data['output'])
 
 if __name__ == '__main__':
     unittest.main()
